# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Simversity API
  version: 0.1.0
  description: AI-powered teaching simulator API for practicing responses to student misconceptions.
  contact:
    name: Simversity
    url: https://simversity.org
  license:
    name: MIT
servers:
  - url: https://simversity.org
    description: Production
  - url: http://localhost:3001
    description: Local development
...
paths:
...
  /api/scenario-builder/chat:
    post:
      tags: [Scenario Builder]
      summary: Chat with AI scenario builder
      description: Streams an AI response that helps the teacher design a scenario. The final assistant message may contain a `<scenario>` JSON block.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [messages]
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    required: [role, content]
                    properties:
                      role:
                        type: string
                        enum: [user, assistant]
                      content:
                        type: string
                        minLength: 1
                        maxLength: 5000
      responses:
        "200":
          description: SSE stream of AI response chunks
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "429":
          $ref: "#/components/responses/RateLimited"
  /api/scenario-builder/create:
    post:
      tags: [Scenario Builder]
      summary: Create course and scenario from builder output
      description: Persists a scenario designed via the builder chat into a new course with personas and agents.
      security:
        - session: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, scenarioTitle, scenarioDescription, students]
              properties:
                subject:
                  type: string
                  minLength: 1
                  maxLength: 200
                gradeLevel:
                  type: string
                  maxLength: 100
                scenarioTitle:
                  type: string
                  minLength: 1
                  maxLength: 300
                scenarioDescription:
                  type: string
                  minLength: 1
                  maxLength: 2000
                activityContext:
                  type: string
                  maxLength: 500
                students:
                  type: array
                  minItems: 1
                  maxItems: 6
                  items:
                    type: object
                    required: [name, description, systemPrompt, openingMessage]
                    properties:
                      name:
                        type: string
                        minLength: 1
                        maxLength: 100
                      description:
                        type: string
                        minLength: 1
                        maxLength: 1000
                      systemPrompt:
                        type: string
                        minLength: 1
                        maxLength: 10000
                      openingMessage:
                        type: string
                        minLength: 1
                        maxLength: 2000
      responses:
        "201":
          description: Scenario created
          content:
            application/json:
              schema:
                type: object
                properties:
                  courseId:
                    type: string
                    format: uuid
                  scenarioId:
                    type: string
                    format: uuid
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
...
  /api/admin/scenarios/{id}:
    get:
      tags: [Admin - Scenarios]
      summary: Get scenario with agents
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Scenario with agents and personas
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Admin - Scenarios]
      summary: Update scenario
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScenarioInput"
      responses:
        "200":
          description: Scenario updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Admin - Scenarios]
      summary: Delete scenario
      description: Fails if conversations reference this scenario.
      security:
        - session: []
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: Scenario deleted
        "400":
          description: Scenario in use by conversations
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: better-auth.session_token
      description: Session cookie set by Better-Auth on sign-in
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ScenarioInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        description:
          type: string
          maxLength: 2000
        observerPrompt:
          type: string
          nullable: true
        activityContext:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        observerModel:
          type: string
          nullable: true
        sortOrder:
          type: integer
        agents:
          type: array
          items:
            type: object
            required: [personaId, openingMessage]
            properties:
              personaId:
                type: string
                format: uuid
              openingMessage:
                type: string
              sortOrder:
                type: integer
...
    SSEEvent:
      description: |
        Server-Sent Event payload. All events use `event: message` with JSON data.
        The `type` field discriminates between event kinds.
      oneOf:
        - type: object
          title: Delta
          description: Streamed AI token chunk
          required: [type, text]
          properties:
            type:
              type: string
              enum: [delta]
            text:
              type: string
        - type: object
          title: Done
          description: Stream complete, response saved to DB
          required: [type, messageId]
          properties:
            type:
              type: string
              enum: [done]
            messageId:
              type: string
              format: uuid
            agentId:
              type: string
              format: uuid
              description: Present on conversation endpoints, absent on observer
            agentName:
              type: string
              description: Present on conversation endpoints, absent on observer
            truncated:
              type: boolean
              description: True when the model hit max_tokens
        - type: object
          title: Error
          description: AI failure or save failure
          required: [type, message]
          properties:
            type:
              type: string
              enum: [error]
            message:
              type: string
      discriminator:
        propertyName: type
...
    Error:
      type: object
      properties:
        error:
          type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"